# Conflicts:
#	src/main/java/com/sh/app/cinema/dto/CinemaDto.java
#	src/main/resources/sql/data.sql

+++ src/main/java/com/sh/app/cinema/entity/Cinema.java
@@ -27,9 +27,9 @@ public class Cinema implements Comparable<Cinema>{
     @Column(nullable = false)
     private String address;
     @Column(nullable = false)
-    private int location_lo; // 지도 경도
+    private Double location_lo; // 지도 경도
     @Column(nullable = false)
-    private int location_la; // 지도 위도
+    private Double location_la; // 지도 위도
     @Column(nullable = false)
     private String phone;
